+++ src/__generated__/ShowHeader_show.graphql.ts
@@ -13,6 +13,7 @@ export type ShowHeader_show = {
     readonly status: string | null;
     readonly partner: ({
         readonly name?: string | null;
+        readonly id?: string;
     }) | null;
     readonly images: ReadonlyArray<({
         readonly url: string | null;
@@ -31,20 +32,24 @@ const node: ConcreteFragment = (function(){
 var v0 = {
   "kind": "ScalarField",
   "alias": null,
-  "name": "__id",
+  "name": "id",
   "args": null,
   "storageKey": null
 },
 v1 = {
   "kind": "ScalarField",
   "alias": null,
-  "name": "name",
+  "name": "__id",
   "args": null,
   "storageKey": null
 },
-v2 = [
-  v1
-];
+v2 = {
+  "kind": "ScalarField",
+  "alias": null,
+  "name": "name",
+  "args": null,
+  "storageKey": null
+};
 return {
   "kind": "Fragment",
   "name": "ShowHeader_show",
@@ -52,15 +57,9 @@ return {
   "metadata": null,
   "argumentDefinitions": [],
   "selections": [
-    {
-      "kind": "ScalarField",
-      "alias": null,
-      "name": "id",
-      "args": null,
-      "storageKey": null
-    },
     v0,
     v1,
+    v2,
     {
       "kind": "ScalarField",
       "alias": null,
@@ -98,16 +97,21 @@ return {
       "concreteType": null,
       "plural": false,
       "selections": [
-        v0,
+        v1,
         {
           "kind": "InlineFragment",
           "type": "ExternalPartner",
-          "selections": v2
+          "selections": [
+            v2
+          ]
         },
         {
           "kind": "InlineFragment",
           "type": "Partner",
-          "selections": v2
+          "selections": [
+            v2,
+            v0
+          ]
         }
       ]
     },
@@ -145,7 +149,7 @@ return {
       "concreteType": "Artist",
       "plural": true,
       "selections": [
-        v1,
+        v2,
         {
           "kind": "ScalarField",
           "alias": null,
@@ -153,11 +157,11 @@ return {
           "args": null,
           "storageKey": null
         },
-        v0
+        v1
       ]
     }
   ]
 };
 })();
-(node as any).hash = '4ccb36c9fbb90a0e85f9d16f4f139549';
+(node as any).hash = '7ad7caebb675a8f3fc7e8c372b7c655c';
 export default node;